Refactor PS and fix minor build issues preventing Travis-CI from completing previously
