+++ src/qt/dynamicgui.cpp
@@ -85,8 +85,8 @@ DynamicGUI::DynamicGUI(const PlatformStyle *_platformStyle, const NetworkStyle *
     labelWalletEncryptionIcon(0),
     clientModel(0),
     walletFrame(0),
-    labelWalletHDStatusIcon(0),
     unitDisplayControl(0),
+    labelWalletHDStatusIcon(0),
     labelConnectionsIcon(0),
     labelBlocksIcon(0),
     progressBarLabel(0),
