+++ src/timedata.h
@@ -10,6 +10,7 @@
 
 #include <algorithm>
 #include <cassert>
+#include <vector>
 
 class CNetAddr;
