+++ src/wallet/wallet.cpp
@@ -22,7 +22,7 @@
 #include "main.h"
 #include "net.h"
 #include "policy/policy.h"
-#include "privatesend.h"
+#include "privatesend-client.h"
 #include "rpcprotocol.h"
 #include "script/script.h"
 #include "script/sign.h"
@@ -1318,7 +1318,7 @@ int CWallet::GetInputPrivateSendRounds(CTxIn txin) const
 {
     LOCK(cs_wallet);
     int realPrivateSendRounds = GetRealInputPrivateSendRounds(txin, 0);
-    return realPrivateSendRounds > nPrivateSendRounds ? nPrivateSendRounds : realPrivateSendRounds;
+    return realPrivateSendRounds > privateSendClient.nPrivateSendRounds ? privateSendClient.nPrivateSendRounds : realPrivateSendRounds;
 }
 
 bool CWallet::IsDenominated(const CTxIn &txin) const
@@ -1958,7 +1958,7 @@ CAmount CWalletTx::GetAnonymizedCredit(bool fUseCache) const
         if(pwallet->IsSpent(hashTx, i) || !pwallet->IsDenominated(txin)) continue;
 
         const int nRounds = pwallet->GetInputPrivateSendRounds(txin);
-        if(nRounds >= nPrivateSendRounds){
+        if(nRounds >= privateSendClient.nPrivateSendRounds){
             nCredit += pwallet->GetCredit(txout, ISMINE_SPENDABLE);
             if (!MoneyRange(nCredit))
                 throw std::runtime_error("CWalletTx::GetAnonymizedCredit() : value out of range");
@@ -2244,7 +2244,7 @@ CAmount CWallet::GetNormalizedAnonymizedBalance() const
                 if (pcoin->GetDepthInMainChain() < 0) continue;
 
                 int nRounds = GetInputPrivateSendRounds(txin);
-                nTotal += pcoin->vout[i].nValue * nRounds / nPrivateSendRounds;
+                nTotal += pcoin->vout[i].nValue * nRounds / privateSendClient.nPrivateSendRounds;
             }
         }
     }
@@ -2257,7 +2257,7 @@ CAmount CWallet::GetNeedsToBeAnonymizedBalance(CAmount nMinBalance) const
     if(fLiteMode) return 0;
 
     CAmount nAnonymizedBalance = GetAnonymizedBalance();
-    CAmount nNeedsToAnonymizeBalance = nPrivateSendAmount*COIN - nAnonymizedBalance;
+    CAmount nNeedsToAnonymizeBalance = privateSendClient.nPrivateSendAmount*COIN - nAnonymizedBalance;
 
     // try to overshoot target PS balance up to nMinBalance
     nNeedsToAnonymizeBalance += nMinBalance;
@@ -2271,7 +2271,7 @@ CAmount CWallet::GetNeedsToBeAnonymizedBalance(CAmount nMinBalance) const
     if(nNeedsToAnonymizeBalance > nAnonymizableBalance) nNeedsToAnonymizeBalance = nAnonymizableBalance;
 
     // we should never exceed the pool max
-    if (nNeedsToAnonymizeBalance > privateSendPool.GetMaxPoolAmount()) nNeedsToAnonymizeBalance = privateSendPool.GetMaxPoolAmount();
+    if (nNeedsToAnonymizeBalance > privateSendClient.GetMaxPoolAmount()) nNeedsToAnonymizeBalance = privateSendClient.GetMaxPoolAmount();
 
     return nNeedsToAnonymizeBalance;
 }
@@ -2644,7 +2644,7 @@ bool CWallet::SelectCoins(const CAmount& nTargetValue, std::set<std::pair<const 
                 CTxIn txin = CTxIn(out.tx->GetHash(),out.i);
                 int nRounds = GetInputPrivateSendRounds(txin);
                 // make sure it's actually anonymized
-                if(nRounds < nPrivateSendRounds) continue;
+                if(nRounds < privateSendClient.nPrivateSendRounds) continue;
             }
             nValueRet += out.tx->vout[out.i].nValue;
             setCoinsRet.insert(std::make_pair(out.tx, out.i));
@@ -2669,7 +2669,7 @@ bool CWallet::SelectCoins(const CAmount& nTargetValue, std::set<std::pair<const 
                         CTxIn txin = CTxIn(out.tx->GetHash(),out.i);
                         int nRounds = GetInputPrivateSendRounds(txin);
                         // make sure it's actually anonymized
-                        if(nRounds < nPrivateSendRounds) continue;
+                        if(nRounds < privateSendClient.nPrivateSendRounds) continue;
                         nValueRet += nDenom;
                         setCoinsRet.insert(std::make_pair(out.tx, out.i));
                     }
@@ -2806,7 +2806,7 @@ bool CWallet::SelectCoinsByDenominations(int nDenom, CAmount nValueMin, CAmount 
     // bit 3 - .1DYN+1
 
     std::vector<int> vecBits;
-    if (!privateSendPool.GetDenominationsBits(nDenom, vecBits)) {
+    if (!privateSendClient.GetDenominationsBits(nDenom, vecBits)) {
         return false;
     }
 
@@ -2903,7 +2903,7 @@ bool CWallet::SelectCoinsGrouppedByAddresses(std::vector<CompactTallyItem>& vecT
                 // otherwise they will just lead to higher fee / lower priority
                 if(wtx.vout[i].nValue <= vecPrivateSendDenominations.back()/10) continue;
                 // ignore anonymized
-                if(GetInputPrivateSendRounds(CTxIn(wtx.GetHash(), i)) >= nPrivateSendRounds) continue;
+                if(GetInputPrivateSendRounds(CTxIn(wtx.GetHash(), i)) >= privateSendClient.nPrivateSendRounds) continue;
             }
 
             CompactTallyItem& item = mapTally[address];
@@ -3355,7 +3355,7 @@ bool CWallet::CreateTransaction(const std::vector<CRecipient>& vecSend, CWalletT
                         nFeeRet += nChange;
                         wtxNew.mapValue["SS"] = "1";
                         // recheck skipped denominations during next mixing
-                        privateSendPool.ClearSkippedDenominations();
+                        privateSendClient.ClearSkippedDenominations();
                     } else {
 
                         // Fill a vout to ourself
@@ -3772,7 +3772,7 @@ bool CWallet::NewKeyPool()
             walletdb.ErasePool(nIndex);
         }
         setExternalKeyPool.clear();
-        fEnablePrivateSend = false;
+        privateSendClient.fEnablePrivateSend = false;
         nKeysLeftSinceAutoBackup = 0;
 
         if (!TopUpKeyPool())
